organised the git in one folder for each topic
the first topic: unemployment duration in M1
the second topic: couple duration, the particular case of conjugal violences in M2
