+++ Couple_duration/Couple_duration_analysis.pdf
